divisor should be a float not int

+++ greencandle/lib/order.py
@@ -84,7 +84,7 @@ class Trade():
             last_open_price = 0
 
         if self.config.main.divisor:
-            proposed_quote_amount = current_quote_bal / int(self.config.main.divisor)
+            proposed_quote_amount = current_quote_bal / float(self.config.main.divisor)
         else:
             proposed_quote_amount = current_quote_bal / (int(self.config.main.max_trades) + 1)
         self.logger.info('item: %s, proposed: %s, last:%s'
